test: move v3 and v5 tests to separate file
